Fix class diagram relationship tap gesture
